GitHub_AI_Agent/requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API Server (for web UI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Data Formats
openpyxl>=3.1.0  # Excel support
pyarrow>=14.0.0  # Parquet support

# Utilities
requests>=2.31.0
